9/29/2019 - Zach Mitchell
--------------------

- Implemented classes and methods for data structures to hold all of the user logged data

--------------------

Next up is to actually layout & connect the UI for adding a list of drops to a log via the Drops tab, and displaying the data for the currently selected boss & individual log via the Logs tab.

I'm also planning to cache some of the Google Sheets data on startup in order to decrease the delta between selecting a boss and the item dropdown being populated in the Drops tab.