10/7/2019 - Zach Mitchell
- Fixed issue in Drops tab where total value text didn't update on removal when the removed item was the last in the list

- Implemented a killcount UI including increment and reset

- Fixed Timer reset button so it stops the timer only instead of switching its state

- Added a ClickBlocker that will be used when dialog, warning, error, input, etc windows pop up so that the user cannot click through them

- Added UI for selecting a log, adding a new log, and updating the currently selected log
	*Added code for only the log dropdown so far: Includes checking for a log data file, creation if none exists, and loading either the default or saved data into cache in DataController
	*Properly works for loading up the program w/ no save data

______________________

Next step will be creating a simple window that will pop up to enter a name for a new log.  After that, I can code that save function and make sure data is properly loaded back in at which point I should be able to move onto the Logs tab which is mostly data display along with basic features (some of which will already be implemented) such as adding a new log, removing a log.