1/8/2020 - Zachary Mitchell
-Switched to a basic event system to reduce a lot of tangling caused by objects needing direct references to each other

-Implemented code to update a selected log
	*Error checking exists in code for bad inputs but is not hooked up to a warning box yet
	*Also needs a check to make sure at least one log exists before allowing an attempt to add to it

-Implemented most of the Logs tab including:
	*Lookup and calculation of data
	*Display and formatting of data to screen
	*Updates of the above as needed when boss or log selection is changed
	*Hookup of the 'Creating a new log' functionality to the button in this tab

___________________________________________________

Next goals are to code the functionality for deleting a log and updating the UI accordingly in the Logs tab.  Then, most likely creation of a confirmation and warning popup for important user decisions and user input problems, respectively.  Almost done with the logging functionality in general (drops+log tab).  Once I have that finished which should just be saving and loading data after what I've already mentioned, I'm planning on doing some QoL improvements, and final error checking and prevention before moving on to the Setup Tab.