1/11/2020 - Zachary Mitchell
- Completed and hooked up a Confirm/Cancel window, Input Warning window

- Added cache of currently selected log for each of the Drops and Logs tabs so that they can be handled individually
	*Currently need to add a check when switching tabs to see is the previously selected log has been deleted

- Added a check to make sure there is a at least one log for the selected boss before trying to add to it

- Fully implemented functionality to delete logs including updating the UI accordingly

- QoL improvements such ability to hit to submit for input fields and ability to tab through relevant interactable UI objects

- UI update so that it is always clear which tab is currently selected

- Small formatting error fixed on Logs UI

- Updated the Google Doc sheet accordingly with semi-recent game update that changed some drops

___________________________________________

So, next up is a quick fix for the error problem noted above (*) in addition to making sure the log dropdowns retain the users choice on each tab independently when adding or removing.  There's some simple UI that I'm going to code to reset when the user updates a log, and a simple check of whether the user wants to exit the program if there are currently drops they have listed but not added to a log.  Those should all be fairly small, so my goal in my next big sitting is to a good amount of the saving and loading log data done.