8/7/2020
-Did a lot of data entry for the RareItemDB

-Cleaned up a little code in my collections and base data classes for items to prepare for the Setup tab requirements

__________________________________________

Going to clean up the existing UI a bit and maybe choose some skins for buttons, inputfields, dropdowns, etc.

After that, I've decided instead of going back and working on back-end code that I know is an improvement (Window stack system, Chain of command for data loading, async, etc) that I will finish up the main functionality first as I'd like to have the entire application working to test and put out publicly.  This is something I'd like to continually work on and possibly support if there's a want for it in the community, so although in other circumstances it might be better to make sure I have an easier foundation to work with, I know I can finish the functionality with what I've started.  Given that and the time since this was first started, I think finishing the forward facing aspects is more important as there will always be new things I learn that could improve back-end systems.