Documentation Updates
Added a brief UI design description and an overview of high level plan for the class designs.

Will be working on that more now, but also working on some Linux/Pentesting stuff in my free time which I may post here if the guide I'm following includes anything that seems fitting to post here.